refactor(civ-layer): split civ_adapter monolith into 5-submodule folder
Splits crates/world-gen/src/civ_adapter.rs (~2450 LOC monolith) into
civ_adapter/ folder with 5 submodules + a re-export shim at mod.rs.
ZERO logic change — diff is pure file shuffling for review readability.

Module layout:

| Module       | LOC | Responsibility |
|--------------|-----|-----------------|
| mesh.rs      | 709 | CivView struct; build_civ_view (flat); ocean augment; sphere projection (project_to_sphere, build_civ_view_spherical); elevation_to_u16; biome / climate translations; private Delaunay + coast helpers; 11 mesh tests |
| pipeline.rs  | 440 | Convenience builders chaining System-A stages (extract_features, build_political, build_settlement, build_routes, build_culture, build_hydrology_view) — all sphere-projected per HIGH-1; 12 pipeline tests including sphere-direction regression |
| naming.rs    | 438 | Synthetic deterministic naming + name pools + LLM-driven name_civ_via_llm via TextProvider; 6 tests using MockTextProvider |
| render.rs    | 385 | Political-map PNG + SVG export + private color/disk/escape helpers + pub(crate) cell_index_to_center for pipeline regression test; 3 render tests |
| bundle.rs    | 220 | CivBundle Serialize/Deserialize + blake3 content_hash + bundle_civ + verify_civ_hash; 4 bundle tests including HIGH-1 unit-vector regression |
| mod.rs       |  47 | pub-use re-export shim preserves crate::civ_adapter::* import surface |

Public API surface unchanged — examples/flatworld.rs and any
downstream consumers compile unchanged.

Total 2239 LOC (saved 206 from removed duplication).

416/416 lib tests pass unchanged (415 active + 1 ignored = live
Postgres). Clippy 3 pre-existing warnings only. Cargo example builds
clean.

Reviewer-friendly module boundaries make PR #9 easier to navigate —
mesh/pipeline/naming/render/bundle each fit on a screen for ~400 LOC
ranges (vs the prior 2450 LOC monolith).
