--- crates/world-gen/src/civ_adapter.rs
@@ -1,2445 +1,0 @@
-//! **Civ Ship 1** — adapter that exposes a `flatworld::FlatWorld` through
-//! the same `(centers, neighbors, biomes, climate, river_flux, is_coast,
-//! elevation, sea_level)` interface the System-A civilization stack
-//! ([`crate::political`], [`crate::settlement`], [`crate::routes`],
-//! [`crate::culture`], [`crate::feature`], [`crate::naming`]) already
-//! consumes. The downstream code is 100% mesh-agnostic — it doesn't care
-//! whether `centers` come from a Fibonacci sphere or a flat rectangle —
-//! so a single adapter unlocks all those layers for the flatworld track.
-//!
-//! ## Cell granularity
-//!
-//! v1 uses **L2 sub-zone = cell** (`world.plates[p].zones[z].subzones[s]`).
-//! The default world is ~150 sub-zones — coarse but enough to validate the
-//! adapter wiring; later civilization ships can opt for finer granularity
-//! by sampling extra cells inside each sub-zone polygon.
-//!
-//! ## Adjacency
-//!
-//! Built via Delaunay triangulation over sub-zone centers using the
-//! `delaunator` crate (already a dependency from `shape::slime`). Each
-//! Delaunay edge becomes a bidirectional neighbor link. This produces a
-//! planar graph with the right asymptotic density for the System-A
-//! political/route algorithms (≈6 neighbors per cell vs. the spherical
-//! Voronoi's ≈6 — same scale).
-//!
-//! ## Biome / climate
-//!
-//! Sub-zone biome is computed by sampling [`flat_climate::compute_zone_climate`]
-//! at the sub-zone centre, then translating the Köppen classification
-//! (`flat_climate::Biome` — 21 cells) to System A's coarser
-//! [`crate::biome::BiomeKind`] (14 cells) and [`crate::climate::ClimateZone`]
-//! (8 cells). The translation is intentionally lossy — System A's
-//! political/settlement code only needs the coarser distinctions to drive
-//! province / settlement / route placement.
-
-use delaunator::{triangulate, Point};
-use serde::{Deserialize, Serialize};
-
-use crate::biome::BiomeKind;
-use crate::climate::ClimateZone;
-use crate::creative_seed::SettlementDensity;
-use crate::culture::{self, Culture};
-use crate::feature::{self, Features};
-use crate::flat_climate::{compute_zone_climate, Biome, WorldClimateParams, ZoneClimate};
-use crate::flatworld::{FlatWorld, BASE_LEVEL};
-use crate::hydrology::{self, Hydrology};
-use crate::political::{self, Political};
-use crate::rng::Rng;
-use crate::routes::{self};
-use crate::settlement::{self};
-use crate::world_map::{CultureRegion, MountainRange, Province, River, Route, Settlement, State, WaterBody};
-
-/// A mesh-shaped view of a `FlatWorld` ready to feed into the System-A
-/// civilization stack. Lengths of every per-cell vector equal
-/// `centers.len()`.
-#[derive(Debug, Clone)]
-pub struct CivView {
-    /// Sub-zone centres in 3D (z=0 for the flat substrate). Cell `i` is
-    /// at `centers[i]`.
-    pub centers: Vec<[f32; 3]>,
-    /// Adjacency lists from a Delaunay triangulation over `centers`. Each
-    /// `neighbors[i]` is the deduped sorted set of cell indices sharing a
-    /// triangle edge with cell `i`.
-    pub neighbors: Vec<Vec<u32>>,
-    /// Coarse System-A biome — translated from the Köppen
-    /// `flat_climate::Biome` via [`koppen_to_biome_kind`].
-    pub biomes: Vec<BiomeKind>,
-    /// Coarse System-A climate zone — translated from `flat_climate`
-    /// per-cell mean temperature + annual precipitation via
-    /// [`derive_climate_zone`].
-    pub climate: Vec<ClimateZone>,
-    /// Per-cell river flux. v1 leaves this at `0.0`; civ-layer Ship 2+
-    /// will sample the flatworld hydrology MVP at the cell centre.
-    pub river_flux: Vec<f32>,
-    /// Whether a cell sits next to the ocean (any neighbor cell is
-    /// `BiomeKind::Ocean`). Used by `settlement::build` to bias toward
-    /// coastal cities.
-    pub is_coast: Vec<bool>,
-    /// Per-cell elevation read from [`FlatWorld::elevation_at`] at the
-    /// sub-zone centre. Units: same as `FlatWorld` elevation (relative;
-    /// `0.0` = void floor, ~1.0 = collision peaks).
-    pub elevation: Vec<f32>,
-    /// Sea level. v1 uses a fixed `0.5` — the FlatWorld's `BASE_LEVEL`
-    /// per-plate floor; cells above are "land", at or below are "sea".
-    pub sea_level: f32,
-}
-
-impl CivView {
-    /// Number of cells.
-    pub fn len(&self) -> usize {
-        self.centers.len()
-    }
-
-    /// True when the cell list is empty (degenerate world).
-    pub fn is_empty(&self) -> bool {
-        self.centers.is_empty()
-    }
-}
-
-/// Build the civilization-view from a `FlatWorld` + the climate params
-/// the caller used (or will use) for biome rendering. The climate sweep
-/// runs **once per sub-zone**, not per pixel — significantly cheaper
-/// than `flat_climate::export_zone_climates`'s default coverage.
-///
-/// **Pre-condition**: `world.plates[*].zones[*].subzones` must be
-/// populated (true since v4.2a SCHEMA ship).
-pub fn build_civ_view(world: &FlatWorld, climate_params: &WorldClimateParams) -> CivView {
-    // Edge-distance grid drives the continentality term in
-    // `compute_zone_climate`. Build it the same way
-    // `flat_climate::export_zone_climates` does: classify every pixel as
-    // sea (no plate covers it) vs. land, then BFS distances out of the
-    // sea set via `zonegen::edge_dist_from_sea`.
-    let w = world.width as usize;
-    let h = world.height as usize;
-    let mut is_sea = vec![false; w * h];
-    for py in 0..h {
-        for px in 0..w {
-            if world
-                .plates_at(px as f32 + 0.5, py as f32 + 0.5)
-                .is_empty()
-            {
-                is_sea[py * w + px] = true;
-            }
-        }
-    }
-    let edge_dist_sea = crate::zonegen::edge_dist_from_sea(&is_sea, w, h);
-
-    let mut centers_2d: Vec<(f32, f32)> = Vec::new();
-    let mut zone_climates: Vec<ZoneClimate> = Vec::new();
-    let mut elevation: Vec<f32> = Vec::new();
-    for plate in &world.plates {
-        for (zi, zone) in plate.zones.iter().enumerate() {
-            // Compute the parent-zone climate once; sub-zones inherit it
-            // for biome/temp/precip in v1. Differentiating climate per
-            // sub-zone is a civ-layer Ship 6+ refinement.
-            let zone_clim = compute_zone_climate(
-                world,
-                climate_params,
-                plate.id,
-                zi,
-                &edge_dist_sea,
-            );
-            for sub in &zone.subzones {
-                centers_2d.push(sub.center);
-                zone_climates.push(zone_clim);
-                elevation.push(world.elevation_at(sub.center.0, sub.center.1));
-            }
-        }
-    }
-
-    let centers: Vec<[f32; 3]> = centers_2d
-        .iter()
-        .map(|&(x, y)| [x, y, 0.0])
-        .collect();
-
-    let neighbors = build_delaunay_neighbors(&centers_2d);
-
-    // Biome derivation: physical features (Ocean / Mountain / Hill) win
-    // over climate-driven Köppen biomes. A sub-zone centre that falls in
-    // void (no plate covers it) is Ocean; high-elevation centres are
-    // Mountain/Hill; everything else is classified by Köppen via
-    // [`koppen_to_biome_kind`].
-    let biomes: Vec<BiomeKind> = centers_2d
-        .iter()
-        .zip(zone_climates.iter())
-        .zip(elevation.iter())
-        .map(|((&(x, y), c), &elev)| {
-            // Thresholds calibrated against `flatworld` elevation
-            // constants: `BASE_LEVEL = 0.35` (interior of a single
-            // plate), collision overlap adds `collision_strength *
-            // collision_gain` (typically 0.05-0.25 per overlapping
-            // pair). Mountain at `> 0.55` catches collision peaks
-            // (≥0.20 above base); Hill at `> 0.45` catches modest
-            // collision uplift (≥0.10 above base). Cells exactly at
-            // BASE_LEVEL stay climate-classified.
-            if world.plates_at(x, y).is_empty() {
-                BiomeKind::Ocean
-            } else if elev > 0.55 {
-                BiomeKind::Mountain
-            } else if elev > 0.45 {
-                BiomeKind::Hill
-            } else {
-                koppen_to_biome_kind(c.biome)
-            }
-        })
-        .collect();
-    let climate: Vec<ClimateZone> = zone_climates
-        .iter()
-        .map(|c| derive_climate_zone(c.temp_mean, c.precip_annual))
-        .collect();
-
-    let is_coast = compute_is_coast(&biomes, &neighbors);
-
-    CivView {
-        centers,
-        neighbors,
-        biomes,
-        climate,
-        river_flux: vec![0.0; centers_2d.len()],
-        is_coast,
-        elevation,
-        sea_level: 0.5,
-    }
-}
-
-/// **Civ Ship 2** — augment a `CivView` with synthetic ocean cells.
-///
-/// Sub-zones are by construction inside plates, so a raw
-/// [`build_civ_view`] never produces `BiomeKind::Ocean` cells. Without
-/// ocean cells:
-///
-/// - [`feature::extract`] can't find water bodies.
-/// - `is_coast` stays false for every cell.
-/// - `settlement::build` has no incentive to place coastal cities.
-///
-/// This pass samples points in the world's **void region** (any spot
-/// where `world.plates_at` returns empty), assigns
-/// `BiomeKind::Ocean` to each, rebuilds the Delaunay adjacency over the
-/// combined point set, and recomputes `is_coast` so coastal land
-/// sub-zones flip to `true`.
-///
-/// Sampling is deterministic — seeded from `world.collision_gain` +
-/// `target_count` so the same `(world, target_count)` always yields the
-/// same ocean point set. Uses simple Poisson-disk rejection with a
-/// minimum-separation derived from the world's area so coverage stays
-/// even.
-pub fn augment_with_ocean(view: CivView, world: &FlatWorld, target_count: usize) -> CivView {
-    let w = world.width as f32;
-    let h = world.height as f32;
-    let target = target_count.max(1);
-
-    // Min-separation tuned for even coverage. Area / target = expected
-    // cell area; the rejection radius is ~70 % of that side length so
-    // we get a slightly-jittered grid pattern, not a Poisson-blue spot
-    // distribution. Coarser is fine — we just need enough ocean cells
-    // for water-body components to form and is_coast to fire.
-    let cell_area = (w * h) / (target as f32);
-    let min_sep = cell_area.sqrt() * 0.7;
-    let min_sep2 = min_sep * min_sep;
-
-    let mut rng = Rng::for_stage(
-        ((world.width as u64) << 32) | (world.height as u64) | (target as u64),
-        b"civ-ocean",
-    );
-    let max_attempts = target.saturating_mul(80).max(2_000);
-    let mut ocean_centers: Vec<(f32, f32)> = Vec::with_capacity(target);
-    let mut attempts = 0_usize;
-    while ocean_centers.len() < target && attempts < max_attempts {
-        attempts += 1;
-        let x = rng.next_f32() * w;
-        let y = rng.next_f32() * h;
-        if !world.plates_at(x, y).is_empty() {
-            continue; // inside a plate — not void
-        }
-        if ocean_centers
-            .iter()
-            .all(|&(ox, oy)| (x - ox) * (x - ox) + (y - oy) * (y - oy) >= min_sep2)
-        {
-            ocean_centers.push((x, y));
-        }
-    }
-
-    if ocean_centers.is_empty() {
-        // World is entirely covered by plates (rare; tiny test worlds).
-        // Return the view unchanged — no synthetic Ocean is honest.
-        return view;
-    }
-
-    // Stitch ocean cells onto the existing CivView.
-    let land_n = view.centers.len();
-    let mut centers_2d: Vec<(f32, f32)> = view
-        .centers
-        .iter()
-        .map(|c| (c[0], c[1]))
-        .collect();
-    let mut centers = view.centers;
-    let mut biomes = view.biomes;
-    let mut climate = view.climate;
-    let mut river_flux = view.river_flux;
-    let mut elevation = view.elevation;
-    for &(ox, oy) in &ocean_centers {
-        centers_2d.push((ox, oy));
-        centers.push([ox, oy, 0.0]);
-        biomes.push(BiomeKind::Ocean);
-        // Ocean climate is `Tropical` (warmest of the system-A buckets);
-        // settlement/route code doesn't read climate on water cells, so
-        // the specific bucket doesn't matter — just stays consistent.
-        climate.push(ClimateZone::Tropical);
-        river_flux.push(0.0);
-        elevation.push(0.0);
-    }
-
-    // Rebuild Delaunay neighbors over the union of land+ocean centers,
-    // then re-derive coast flags on the joint biome vec.
-    let neighbors = build_delaunay_neighbors(&centers_2d);
-    let is_coast = compute_is_coast(&biomes, &neighbors);
-
-    debug_assert_eq!(centers.len(), land_n + ocean_centers.len());
-    debug_assert_eq!(biomes.len(), centers.len());
-    debug_assert_eq!(neighbors.len(), centers.len());
-
-    CivView {
-        centers,
-        neighbors,
-        biomes,
-        climate,
-        river_flux,
-        is_coast,
-        elevation,
-        sea_level: view.sea_level,
-    }
-}
-
-/// **Civ Ship 2** — convenience pipeline: build a civ view, augment it
-/// with synthetic ocean cells, then run System-A's mesh-agnostic
-/// [`feature::extract`] to get named mountain ranges, rivers, and water
-/// bodies for downstream naming / political / settlement layers.
-///
-/// `ocean_target` is the number of ocean cells to synthesize. A default
-/// world (1280×720, 12 plates) reads well with 40-80 ocean cells; tiny
-/// test worlds with 8-16. Pass `0` to skip ocean augmentation (matches
-/// Ship-1 behaviour).
-///
-/// Returns `(view, features)` so callers can pass both to System-A's
-/// political / settlement / route builders without re-running the
-/// expensive climate sweep inside [`build_civ_view`].
-pub fn extract_features(
-    world: &FlatWorld,
-    climate_params: &WorldClimateParams,
-    ocean_target: usize,
-) -> (CivView, Features) {
-    let view = build_civ_view(world, climate_params);
-    let mut view = if ocean_target > 0 {
-        augment_with_ocean(view, world, ocean_target)
-    } else {
-        view
-    };
-    let features = feature::extract(&view.biomes, &view.neighbors);
-    // **review-impl HIGH-1**: System-A's civilization builders
-    // (political / settlement / routes / culture / pathfind::spaced_ok)
-    // use a *sphere-specific* distance metric `1 - dot(c_a, c_b)` that
-    // is only meaningful when `centers` are unit vectors. On flat
-    // substrate (`z = 0`, `x, y ∈ [0, w]`) the dot product is hugely
-    // positive and the metric flips sign + scale — silently producing
-    // degenerate seed placement (province seeds cluster in the
-    // smallest-x/smallest-y corner; settlement-capital matching breaks
-    // similarly). Projecting *after* Delaunay (already built on the
-    // flat plane) preserves topology while giving the downstream code
-    // the unit-vector centres it was written for.
-    project_to_sphere(&mut view, world);
-    (view, features)
-}
-
-/// **Civ Ship 4** — quantise the civ view's f32 elevation into the
-/// u16 grid System-A's hydrology stack expects. Returns the per-cell
-/// `Vec<u16>` and the matching `sea_level` cutoff.
-///
-/// Calibration: flatworld f32 elevation lives in `[0.0, ~0.9]` —
-/// `0.0` is void / synthetic ocean, `BASE_LEVEL = 0.35` is the
-/// interior of an uncontested plate, and collision overlap can push
-/// land cells up toward ~0.9. We linearly map `[0, 1]` → `[0, 65535]`
-/// (clamping past `1.0`) and set `sea_level_u16` halfway between
-/// void (`0`) and `BASE_LEVEL` so every void / synthetic-ocean cell
-/// (elev `0.0`) lands strictly below sea level and every plate
-/// interior (`≥ BASE_LEVEL`) lands strictly above.
-pub fn elevation_to_u16(view: &CivView) -> (Vec<u16>, u16) {
-    let elev: Vec<u16> = view
-        .elevation
-        .iter()
-        .map(|&e| (e.clamp(0.0, 1.0) * 65535.0).round() as u16)
-        .collect();
-    let sea_level = ((BASE_LEVEL * 0.5) * 65535.0).round() as u16;
-    (elev, sea_level)
-}
-
-/// **Civ Ship 4** — run System-A's hydrology pipeline (priority-flood
-/// receiver graph + flow accumulation → river flux) on the civ view.
-/// The output's `river_flux` and `is_coast` overrides the view's
-/// adjacency-only is_coast from Ship 2 — hydrology's coast detection
-/// uses real connectivity to the ocean network, not just immediate
-/// neighbors.
-pub fn build_hydrology_view(view: &CivView) -> Hydrology {
-    let (elev_u16, sea_level_u16) = elevation_to_u16(view);
-    hydrology::build(
-        &view.centers,
-        &elev_u16,
-        sea_level_u16,
-        &view.neighbors,
-        &view.climate,
-    )
-}
-
-/// **Civ Ship 4** — full pipeline through System-A's settlement
-/// builder. Chains build_political → hydrology → settlement::build so
-/// the caller gets the full (CivView, Features, Political, Hydrology,
-/// Vec<Settlement>) bundle. Hydrology output overrides the view's
-/// `river_flux` and `is_coast` since it computes both from real
-/// drainage analysis.
-pub fn build_settlement(
-    world: &FlatWorld,
-    climate_params: &WorldClimateParams,
-    ocean_target: usize,
-    seed: u64,
-    density: SettlementDensity,
-) -> (CivView, Features, Political, Hydrology, Vec<Settlement>) {
-    let (mut view, features, political) =
-        build_political(world, climate_params, ocean_target, seed);
-    let hydro = build_hydrology_view(&view);
-    view.river_flux = hydro.river_flux.clone();
-    view.is_coast = hydro.is_coast.clone();
-    let settlements = settlement::build(
-        seed,
-        &view.centers,
-        &view.biomes,
-        &view.climate,
-        &view.river_flux,
-        &view.is_coast,
-        density,
-        &political,
-    );
-    (view, features, political, hydro, settlements)
-}
-
-// ============================================================================
-// Civ Approach A step 2 — sphere mesh adapter (3D upgrade)
-// ============================================================================
-
-/// **Approach A step 2** — project a flat civ view onto a unit sphere
-/// via equirectangular projection.
-///
-/// For each cell centre `(x, y, 0)` in `view.centers`:
-/// - longitude `lon = (x / w − 0.5) · 2π ∈ [−π, π]`
-/// - latitude  `lat = (0.5 − y / h) · π ∈ [−π/2, π/2]` (y=0 is north pole)
-/// - 3D unit sphere point `(cos(lat)·cos(lon), cos(lat)·sin(lon), sin(lat))`
-///
-/// **Topology caveat (LOW-1).** Equirectangular is a homeomorphism only
-/// on the *open* rectangle interior: the left and right edges
-/// (`x = 0` and `x = w`) map to the **same meridian** (the
-/// antimeridian); the top and bottom edges (`y = 0` and `y = h`) map
-/// to **single points** (the poles). Plate polygons that touch any
-/// boundary edge will have Delaunay neighbour links that no longer
-/// represent adjacency on the sphere. Phase-A continent-scale worlds
-/// keep plates well inside the interior via `flatworld::place_centers`'s
-/// `min_sep` margin, so this is empirically fine today; future
-/// `FlatParams` tunings that allow boundary-touching plates will need
-/// either a different projection or a true spherical Delaunay.
-///
-/// **Why this projection is the correct substrate (not just an
-/// upgrade).** System-A's civilization builders (`political::build` at
-/// `political.rs:185`, `settlement::build` at `settlement.rs:121`,
-/// `routes::build` at `routes.rs:183`, `pathfind::spaced_ok`) compute
-/// distance as `1 - dot(c_a, c_b)` — a formula that is
-/// monotone-equivalent to great-circle distance only when both
-/// `c_a` and `c_b` are **unit vectors**. On flat `[x, y, 0]` centres
-/// with `x, y ∈ [0, w]` the dot product is huge and positive and the
-/// metric becomes degenerate (see HIGH-1 in the 2026-05-30
-/// /review-impl report). The downstream code path is shared between
-/// flat and sphere views, but its output is only semantically correct
-/// on the sphere view.
-pub fn project_to_sphere(view: &mut CivView, world: &FlatWorld) {
-    let w = world.width as f32;
-    let h = world.height as f32;
-    if w <= 0.0 || h <= 0.0 {
-        return;
-    }
-    for c in view.centers.iter_mut() {
-        let (x, y) = (c[0], c[1]);
-        let lon = (x / w - 0.5) * std::f32::consts::TAU;
-        let lat = (0.5 - y / h) * std::f32::consts::PI;
-        let cos_lat = lat.cos();
-        c[0] = cos_lat * lon.cos();
-        c[1] = cos_lat * lon.sin();
-        c[2] = lat.sin();
-    }
-}
-
-/// **Approach A step 2** — explicit-sphere variant of [`build_civ_view`]
-/// that bakes the [`project_to_sphere`] step in.
-///
-/// Most callers want this. [`build_civ_view`] is kept as a flat opt-out
-/// for the Ship-1 Delaunay tests that need direct access to the planar
-/// adjacency; the `extract_features` /
-/// [`build_political`] / [`build_settlement`] / [`build_routes`] /
-/// [`build_culture`] / [`bundle_civ`] convenience functions all
-/// internally project to sphere (see HIGH-1 in /review-impl) so their
-/// outputs match the metric assumed by System-A's civilization
-/// builders.
-pub fn build_civ_view_spherical(
-    world: &FlatWorld,
-    climate_params: &WorldClimateParams,
-) -> CivView {
-    let mut view = build_civ_view(world, climate_params);
-    project_to_sphere(&mut view, world);
-    view
-}
-
-// ============================================================================
-// Civ Ship 7b — LLM-driven naming via TextProvider trait
-// ============================================================================
-
-use crate::shape::llm::{LlmError, TextPrompt, TextProvider};
-
-/// **Civ Ship 7b** — JSON schema the LLM-driven naming caller asks the
-/// [`TextProvider`] to fill. One string array per named-feature
-/// category. Matches the structure System-A's `naming::WorldNames`
-/// expects, kept local here so we don't need to make the original `pub`.
-fn civ_names_schema() -> serde_json::Value {
-    let str_array = || serde_json::json!({ "type": "array", "items": { "type": "string" } });
-    serde_json::json!({
-        "type": "object",
-        "additionalProperties": false,
-        "required": [
-            "settlements", "states", "provinces", "cultures",
-            "mountain_ranges", "rivers", "water_bodies"
-        ],
-        "properties": {
-            "settlements": str_array(),
-            "states": str_array(),
-            "provinces": str_array(),
-            "cultures": str_array(),
-            "mountain_ranges": str_array(),
-            "rivers": str_array(),
-            "water_bodies": str_array(),
-        }
-    })
-}
-
-/// Decoded structure of the LLM's reply. Each field is
-/// `#[serde(default)]` so an omitted category leaves those features
-/// unnamed rather than failing the whole parse — matches System A's
-/// `naming::WorldNames` behaviour.
-#[derive(Debug, serde::Deserialize)]
-struct CivNames {
-    #[serde(default)]
-    settlements: Vec<String>,
-    #[serde(default)]
-    states: Vec<String>,
-    #[serde(default)]
-    provinces: Vec<String>,
-    #[serde(default)]
-    cultures: Vec<String>,
-    #[serde(default)]
-    mountain_ranges: Vec<String>,
-    #[serde(default)]
-    rivers: Vec<String>,
-    #[serde(default)]
-    water_bodies: Vec<String>,
-}
-
-/// **Civ Ship 7b** — name civilization features via an
-/// [`TextProvider`]. Builds a structured prompt with feature counts,
-/// asks for one name list per category, parses the JSON reply, applies
-/// the names by position (short lists leave surplus features unnamed;
-/// long lists are truncated).
-///
-/// **NOT deterministic when using a real LLM.** [`MockTextProvider`]
-/// IS deterministic; Anthropic / OpenAI / Ollama at `temperature = 0`
-/// are deterministic-ish (sometimes drift on retries). Treat this as an
-/// authoring step separate from the deterministic Ship 7
-/// [`apply_synthetic_names`] — use one or the other per render.
-pub fn name_civ_via_llm(
-    features: &mut Features,
-    political: &mut Political,
-    settlements: &mut [Settlement],
-    culture: &mut Culture,
-    provider: &dyn TextProvider,
-    archetype: &str,
-) -> Result<(), LlmError> {
-    let system = "You are a world-naming assistant for a procedural fantasy-map \
-                  generator. Given a world's archetype and the counts of its \
-                  features, produce ONE JSON object of name lists matching the \
-                  provided schema. For each category produce exactly the \
-                  requested number of distinct, evocative proper names that \
-                  fit the archetype. Output only the JSON object.";
-    let user = format!(
-        "Archetype: {archetype}.\n\
-         Name these features — produce exactly the given count of distinct \
-         names per category:\n\
-         - settlements: {}\n\
-         - states: {}\n\
-         - provinces: {}\n\
-         - cultures: {}\n\
-         - mountain_ranges: {}\n\
-         - rivers: {}\n\
-         - water_bodies: {}",
-        settlements.len(),
-        political.states.len(),
-        political.provinces.len(),
-        culture.culture_regions.len(),
-        features.mountain_ranges.len(),
-        features.rivers.len(),
-        features.water_bodies.len(),
-    );
-
-    let prompt = TextPrompt::new(system, user).with_schema(civ_names_schema(), "world_names");
-    let raw = provider.complete(&prompt)?;
-    let names: CivNames = serde_json::from_str(raw.trim()).map_err(|e| {
-        LlmError::InvalidResponse(format!("CivNames JSON parse failed: {e}"))
-    })?;
-
-    for (s, n) in settlements.iter_mut().zip(&names.settlements) {
-        s.name = n.clone();
-    }
-    for (s, n) in political.states.iter_mut().zip(&names.states) {
-        s.name = n.clone();
-    }
-    for (p, n) in political.provinces.iter_mut().zip(&names.provinces) {
-        p.name = n.clone();
-    }
-    for (c, n) in culture.culture_regions.iter_mut().zip(&names.cultures) {
-        c.name = n.clone();
-    }
-    for (mr, n) in features.mountain_ranges.iter_mut().zip(&names.mountain_ranges) {
-        mr.name = n.clone();
-    }
-    for (rv, n) in features.rivers.iter_mut().zip(&names.rivers) {
-        rv.name = n.clone();
-    }
-    for (wb, n) in features.water_bodies.iter_mut().zip(&names.water_bodies) {
-        wb.name = n.clone();
-    }
-    Ok(())
-}
-
-// ============================================================================
-// Civ Ship 9 — CivBundle + JSON round-trip + content_hash
-// ============================================================================
-
-/// **Civ Ship 9** — flat, Serialize-able bundle of every civ-layer
-/// output. Use [`bundle_civ`] to build one from the full pipeline; use
-/// [`compute_civ_hash`] to (re)verify the bundled state.
-///
-/// All embedded types come from `world_map.rs` (Province / State /
-/// Settlement / Route / CultureRegion / MountainRange / River /
-/// WaterBody / BiomeKind / ClimateZone), which already derive
-/// Serialize / Deserialize — so JSON round-trip is automatic. Fields
-/// from Political / Hydrology / Culture are unrolled to their component
-/// vectors so we don't have to add Serialize derives to System-A's
-/// stage-output structs (would ripple through `lib.rs::generate`).
-///
-/// `content_hash` is blake3 over every field except itself, so a
-/// freshly-built bundle that gets serialized → deserialized must
-/// validate via [`verify_civ_hash`].
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CivBundle {
-    pub centers: Vec<[f32; 3]>,
-    pub neighbors: Vec<Vec<u32>>,
-    pub biomes: Vec<BiomeKind>,
-    pub climate: Vec<ClimateZone>,
-    pub river_flux: Vec<f32>,
-    pub is_coast: Vec<bool>,
-    pub elevation: Vec<f32>,
-    pub sea_level: f32,
-
-    // From Political (unrolled).
-    pub province_of: Vec<u32>,
-    pub provinces: Vec<Province>,
-    pub states: Vec<State>,
-
-    pub settlements: Vec<Settlement>,
-    pub routes: Vec<Route>,
-
-    // From Culture (unrolled).
-    pub culture_of: Vec<u32>,
-    pub culture_regions: Vec<CultureRegion>,
-
-    // From Features (unrolled).
-    pub mountain_ranges: Vec<MountainRange>,
-    pub rivers: Vec<River>,
-    pub water_bodies: Vec<WaterBody>,
-
-    /// Blake3 digest over every other field. Computed by
-    /// [`bundle_civ`]; cross-check with [`verify_civ_hash`] after
-    /// JSON round-trip.
-    pub content_hash: [u8; 32],
-}
-
-/// **Civ Ship 9** — run the full civ pipeline (build_culture +
-/// apply_synthetic_names) and pack the output into a [`CivBundle`]
-/// with a fresh [`content_hash`].
-#[allow(clippy::too_many_arguments)]
-pub fn bundle_civ(
-    world: &FlatWorld,
-    climate_params: &WorldClimateParams,
-    ocean_target: usize,
-    seed: u64,
-    density: SettlementDensity,
-    culture_count: u8,
-) -> CivBundle {
-    let (view, mut features, mut political, _hydro, mut settlements, routes_v, mut culture_v) =
-        build_culture(world, climate_params, ocean_target, seed, density, culture_count);
-    apply_synthetic_names(
-        &mut features,
-        &mut political,
-        &mut settlements,
-        &mut culture_v,
-        seed,
-    );
-    let mut bundle = CivBundle {
-        centers: view.centers,
-        neighbors: view.neighbors,
-        biomes: view.biomes,
-        climate: view.climate,
-        river_flux: view.river_flux,
-        is_coast: view.is_coast,
-        elevation: view.elevation,
-        sea_level: view.sea_level,
-        province_of: political.province_of,
-        provinces: political.provinces,
-        states: political.states,
-        settlements,
-        routes: routes_v,
-        culture_of: culture_v.culture_of,
-        culture_regions: culture_v.culture_regions,
-        mountain_ranges: features.mountain_ranges,
-        rivers: features.rivers,
-        water_bodies: features.water_bodies,
-        content_hash: [0u8; 32],
-    };
-    bundle.content_hash = compute_civ_hash(&bundle);
-    bundle
-}
-
-/// Blake3 digest of every CivBundle field EXCEPT `content_hash` itself.
-///
-/// Implementation: clone the bundle with `content_hash = [0; 32]`,
-/// serialize the clone to JSON, hash the bytes. `serde_json` output is
-/// stable for a given struct definition, the encoding is already a
-/// dependency, and the round-trip test asserts the only property that
-/// matters externally — deserialize-then-rehash equals the original
-/// hash. Switch to a `bincode`-style fixed encoder later if hash
-/// computation becomes a hot path.
-pub fn compute_civ_hash(bundle: &CivBundle) -> [u8; 32] {
-    let mut hasher = blake3::Hasher::new();
-    // The simplest stable encoding: serialize to JSON without
-    // content_hash, then hash that. JSON serialization is already a
-    // dependency of the crate, the output is stable for `serde_json`,
-    // and the round-trip test asserts exactly the property that
-    // matters (deserialize-then-rehash equals the original hash).
-    let mut clone = bundle.clone();
-    clone.content_hash = [0u8; 32];
-    let bytes = serde_json::to_vec(&clone)
-        .expect("CivBundle is composed of Serialize types; encoding cannot fail");
-    hasher.update(&bytes);
-    *hasher.finalize().as_bytes()
-}
-
-/// Recompute the hash and compare it to the bundled `content_hash`.
-/// Used by the round-trip test to assert serialize → deserialize did
-/// not perturb any field.
-pub fn verify_civ_hash(bundle: &CivBundle) -> bool {
-    compute_civ_hash(bundle) == bundle.content_hash
-}
-
-// ============================================================================
-// Civ Ship 8 — political PNG + SVG export
-// ============================================================================
-
-/// **Civ Ship 8** — build a `HashMap` from `(plate_id, zone_id,
-/// subzone_id)` to the civ-layer cell index that the political /
-/// settlement / hydrology vectors are keyed by. Iteration order matches
-/// [`build_civ_view`] — plates outer, zones middle, subzones inner —
-/// so the mapping is a stable function of `world.plates`.
-fn cell_index_map(world: &FlatWorld) -> std::collections::HashMap<(usize, usize, usize), u32> {
-    let mut out = std::collections::HashMap::new();
-    let mut idx: u32 = 0;
-    for plate in &world.plates {
-        for (zi, zone) in plate.zones.iter().enumerate() {
-            for si in 0..zone.subzones.len() {
-                out.insert((plate.id, zi, si), idx);
-                idx += 1;
-            }
-        }
-    }
-    out
-}
-
-/// Deterministic per-state colour. Stays stable across renders so the
-/// political map can be compared frame-to-frame.
-fn state_color(state_id: u32) -> [u8; 3] {
-    let hue = (state_id.wrapping_mul(57) % 360) as f32;
-    let sat = 0.55;
-    let val = 0.85;
-    hsv_to_rgb(hue, sat, val)
-}
-
-/// Simple HSV → RGB helper (saturation+value in `[0, 1]`, hue in degrees).
-fn hsv_to_rgb(h: f32, s: f32, v: f32) -> [u8; 3] {
-    let c = v * s;
-    let hp = h / 60.0;
-    let x = c * (1.0 - (hp % 2.0 - 1.0).abs());
-    let (r, g, b) = match hp as u32 {
-        0 => (c, x, 0.0),
-        1 => (x, c, 0.0),
-        2 => (0.0, c, x),
-        3 => (0.0, x, c),
-        4 => (x, 0.0, c),
-        _ => (c, 0.0, x),
-    };
-    let m = v - c;
-    [
-        ((r + m) * 255.0).round() as u8,
-        ((g + m) * 255.0).round() as u8,
-        ((b + m) * 255.0).round() as u8,
-    ]
-}
-
-/// **Civ Ship 8** — political-map RGB buffer.
-///
-/// For each pixel: find the covering plate → its sub-zone via
-/// [`crate::flatworld::Plate::zone_at_polygon`] +
-/// [`crate::flatworld::Zone::subzone_at_polygon`] → look up
-/// `political.province_of[cell] → political.provinces[…].state` → colour
-/// by `state_color`. Void pixels render as dark ocean blue; settlement
-/// cells get a white dot overlay.
-///
-/// Buffer layout: `width * height * 3` RGB bytes, top-left origin.
-pub fn render_civ_political_png(
-    world: &FlatWorld,
-    political: &Political,
-    settlements: &[Settlement],
-) -> Vec<u8> {
-    const VOID: [u8; 3] = [12, 22, 48]; // deep-ocean blue
-    let w = world.width as usize;
-    let h = world.height as usize;
-    let mut buf = vec![0u8; w * h * 3];
-    let idx_of = cell_index_map(world);
-
-    for py in 0..h {
-        for px in 0..w {
-            let fx = px as f32 + 0.5;
-            let fy = py as f32 + 0.5;
-            let pixel_color = match world.plates.iter().find(|p| p.contains(fx, fy)) {
-                None => VOID,
-                Some(plate) => {
-                    let zi_opt = plate.zone_at_polygon(fx, fy);
-                    let mut color = VOID;
-                    if let Some(zi) = zi_opt {
-                        let zone = &plate.zones[zi];
-                        if let Some(si) = zone.subzone_at_polygon(fx, fy) {
-                            if let Some(&cell_idx) = idx_of.get(&(plate.id, zi, si)) {
-                                let prov = political.province_of[cell_idx as usize];
-                                if prov != u32::MAX {
-                                    let state = political.provinces[prov as usize].state;
-                                    color = state_color(state);
-                                }
-                            }
-                        }
-                    }
-                    color
-                }
-            };
-            let off = (py * w + px) * 3;
-            buf[off] = pixel_color[0];
-            buf[off + 1] = pixel_color[1];
-            buf[off + 2] = pixel_color[2];
-        }
-    }
-
-    // Settlement dots — radius 3 px filled white.
-    for s in settlements {
-        let cell = s.cell as usize;
-        // Reverse-lookup centre from cell index by re-walking the
-        // plate/zone/subzone tree. Cheap because settlements is small.
-        if let Some(center) = cell_index_to_center(world, cell) {
-            paint_disk(&mut buf, w, h, center.0 as i32, center.1 as i32, 3, [255, 255, 255]);
-        }
-    }
-
-    buf
-}
-
-fn cell_index_to_center(world: &FlatWorld, target: usize) -> Option<(f32, f32)> {
-    let mut idx: usize = 0;
-    for plate in &world.plates {
-        for zone in &plate.zones {
-            for sub in &zone.subzones {
-                if idx == target {
-                    return Some(sub.center);
-                }
-                idx += 1;
-            }
-        }
-    }
-    None
-}
-
-fn paint_disk(buf: &mut [u8], w: usize, h: usize, cx: i32, cy: i32, r: i32, color: [u8; 3]) {
-    let r2 = r * r;
-    for dy in -r..=r {
-        for dx in -r..=r {
-            if dx * dx + dy * dy > r2 {
-                continue;
-            }
-            let x = cx + dx;
-            let y = cy + dy;
-            if x < 0 || y < 0 || (x as usize) >= w || (y as usize) >= h {
-                continue;
-            }
-            let off = (y as usize * w + x as usize) * 3;
-            buf[off] = color[0];
-            buf[off + 1] = color[1];
-            buf[off + 2] = color[2];
-        }
-    }
-}
-
-/// **Civ Ship 8** — SVG export of the civ bundle.
-///
-/// Layers (back to front):
-/// 1. World background rectangle (light ocean blue).
-/// 2. Plate polygons filled with their state hue (light alpha).
-/// 3. Route polylines connecting cell centres along each route's path.
-/// 4. Settlement circles + text labels with their synthetic name.
-/// 5. Province name labels at the capital cell centre.
-///
-/// Designed for diff-friendly text output; not optimised for visual
-/// fidelity. PNG remains the production render.
-#[allow(clippy::too_many_arguments)]
-pub fn render_civ_svg(
-    world: &FlatWorld,
-    political: &Political,
-    settlements: &[Settlement],
-    routes_v: &[Route],
-) -> String {
-    let w = world.width;
-    let h = world.height;
-    let mut svg = String::new();
-    svg.push_str(&format!(
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n\
-         <svg xmlns=\"http://www.w3.org/2000/svg\" viewBox=\"0 0 {w} {h}\" \
-         width=\"{w}\" height=\"{h}\">\n",
-    ));
-    svg.push_str(&format!(
-        "  <rect x=\"0\" y=\"0\" width=\"{w}\" height=\"{h}\" fill=\"#0c1630\"/>\n",
-    ));
-
-    // Plates filled with their primary subzone's state colour.
-    let idx_of = cell_index_map(world);
-    for plate in &world.plates {
-        // Use the first sub-zone's state as the plate "tint" — good
-        // enough for an at-a-glance view.
-        let tint = plate
-            .zones
-            .iter()
-            .enumerate()
-            .find_map(|(zi, zone)| {
-                zone.subzones.iter().enumerate().find_map(|(si, _)| {
-                    idx_of.get(&(plate.id, zi, si)).copied()
-                })
-            })
-            .map(|cell| {
-                let prov = political.province_of[cell as usize];
-                if prov == u32::MAX {
-                    [128, 128, 160]
-                } else {
-                    state_color(political.provinces[prov as usize].state)
-                }
-            })
-            .unwrap_or([128, 128, 160]);
-        let fill = format!("#{:02x}{:02x}{:02x}", tint[0], tint[1], tint[2]);
-        for poly in &plate.components {
-            let points = poly
-                .iter()
-                .map(|&(x, y)| format!("{:.1},{:.1}", x, y))
-                .collect::<Vec<_>>()
-                .join(" ");
-            svg.push_str(&format!(
-                "  <polygon points=\"{points}\" fill=\"{fill}\" \
-                 fill-opacity=\"0.5\" stroke=\"#222\" stroke-width=\"1\"/>\n",
-            ));
-        }
-    }
-
-    // Routes — emit polylines along each route's `path` (cell index
-    // list). Skip zero-hop or path-less routes.
-    for r in routes_v {
-        if r.path.len() < 2 {
-            continue;
-        }
-        let pts: Vec<(f32, f32)> = r
-            .path
-            .iter()
-            .filter_map(|&c| cell_index_to_center(world, c as usize))
-            .collect();
-        if pts.len() < 2 {
-            continue;
-        }
-        let coords = pts
-            .iter()
-            .map(|&(x, y)| format!("{:.1},{:.1}", x, y))
-            .collect::<Vec<_>>()
-            .join(" ");
-        let stroke = match r.kind {
-            crate::world_map::RouteKind::Road => "#3a2a18",
-            crate::world_map::RouteKind::SeaLane => "#1c4a7a",
-            crate::world_map::RouteKind::Trail => "#5a4a2a",
-            crate::world_map::RouteKind::RiverNavigation => "#2a8aa0",
-            crate::world_map::RouteKind::MountainPass => "#7a6a5a",
-        };
-        svg.push_str(&format!(
-            "  <polyline points=\"{coords}\" fill=\"none\" stroke=\"{stroke}\" \
-             stroke-width=\"1.5\" stroke-opacity=\"0.85\"/>\n",
-        ));
-    }
-
-    // Settlements + labels.
-    for s in settlements {
-        if let Some((x, y)) = cell_index_to_center(world, s.cell as usize) {
-            svg.push_str(&format!(
-                "  <circle cx=\"{:.1}\" cy=\"{:.1}\" r=\"3\" fill=\"#fff\" stroke=\"#222\"/>\n",
-                x, y
-            ));
-            let escaped = escape_xml(&s.name);
-            svg.push_str(&format!(
-                "  <text x=\"{:.1}\" y=\"{:.1}\" font-size=\"10\" fill=\"#fff\" \
-                 stroke=\"#000\" stroke-width=\"0.4\" \
-                 font-family=\"sans-serif\">{escaped}</text>\n",
-                x + 4.0,
-                y - 4.0
-            ));
-        }
-    }
-
-    // Province capital labels.
-    for prov in &political.provinces {
-        if let Some((x, y)) = cell_index_to_center(world, prov.capital_cell as usize) {
-            let escaped = escape_xml(&prov.name);
-            svg.push_str(&format!(
-                "  <text x=\"{:.1}\" y=\"{:.1}\" font-size=\"8\" fill=\"#eee\" \
-                 fill-opacity=\"0.85\" font-style=\"italic\" \
-                 font-family=\"serif\">{escaped}</text>\n",
-                x + 5.0,
-                y + 10.0
-            ));
-        }
-    }
-
-    svg.push_str("</svg>\n");
-    svg
-}
-
-fn escape_xml(s: &str) -> String {
-    s.replace('&', "&amp;")
-        .replace('<', "&lt;")
-        .replace('>', "&gt;")
-        .replace('"', "&quot;")
-        .replace('\'', "&apos;")
-}
-
-// ============================================================================
-// Civ Ship 7 — synthetic deterministic naming
-// ============================================================================
-
-// Pools of evocative fantasy names — seeded RNG picks one per feature.
-// Short on purpose: the goal of Ship 7 is to give every feature a
-// stable, non-empty name so SVG export (Ship 9) has labels to render.
-// LLM-driven naming via a future `TextProvider` trait can replace these
-// in Ship 7b without changing the rest of the civ pipeline.
-const SETTLEMENT_NAMES: &[&str] = &[
-    "Aetherholt", "Brightford", "Cinderwatch", "Dawnreach", "Embervale",
-    "Frostmere", "Goldenhall", "Hollowbarrow", "Ironkeep", "Jadewood",
-    "Kingsford", "Larkspur", "Mistmoor", "Northwatch", "Oakenshield",
-    "Pinevale", "Quietwater", "Ravenholm", "Stonehearth", "Thornbury",
-    "Umbergate", "Vinewreath", "Willowbrook", "Yewglen", "Zephyrport",
-];
-
-const STATE_NAMES: &[&str] = &[
-    "Aelvarra", "Brennor", "Caldaris", "Drakhalim", "Eronthel",
-    "Faerondale", "Glenwarde", "Hjarsgrad", "Iskandar", "Jorvik",
-    "Kelmarine", "Lirenoth", "Mythraal", "Northkin", "Ostralia",
-    "Parthenor", "Querion", "Rikhalim", "Sundarial", "Thalassia",
-];
-
-const PROVINCE_PREFIXES: &[&str] = &[
-    "Vale of", "March of", "Reach of", "Hold of", "Span of", "Realm of",
-    "Domain of", "Land of",
-];
-
-const PROVINCE_ROOTS: &[&str] = &[
-    "Ashwynne", "Briarfall", "Caelwood", "Deepford", "Elder Pines",
-    "Falconcrest", "Glimmerlake", "Hawthorn", "Ironbark", "Larkmere",
-    "Mistgate", "Nightspire", "Oldhollow", "Pondbridge", "Quartzcliff",
-    "Redmoor", "Silverbrook", "Twilight Glade", "Umbra", "Whitewater",
-];
-
-const CULTURE_NAMES: &[&str] = &[
-    "Aelir", "Brenn", "Caelori", "Dhuran", "Eldari",
-    "Fenni", "Gwynar", "Hjorl", "Iskari", "Jolvik",
-    "Kelmar", "Lirthen", "Myrran", "Norval", "Ostren",
-];
-
-const MOUNTAIN_DESCRIPTORS: &[&str] = &[
-    "Cloudpiercer", "Frostfang", "Sunspear", "Stormcrown", "Ashpeak",
-    "Greyhorn", "Ironreach", "Mistwall", "Skyforge", "Thunderridge",
-];
-
-const RIVER_DESCRIPTORS: &[&str] = &[
-    "Quickwater", "Silvercourse", "Black", "Goldrun", "Whisper",
-    "Coldstream", "Greenway", "Bright", "Stillrun", "Hollow",
-];
-
-const WATER_BODY_DESCRIPTORS: &[&str] = &[
-    "Whitecap Sea", "Sunken Bay", "Twilight Reach", "Sapphire Strait",
-    "Mistwarden Sea", "Forgotten Bay", "Crystal Reach", "Halcyon Sound",
-    "Verdant Coast", "Stormwatch Sea",
-];
-
-fn pick<'a>(pool: &'a [&'a str], rng: &mut Rng) -> &'a str {
-    pool[(rng.next_u32() as usize) % pool.len()]
-}
-
-/// **Civ Ship 7** — assign deterministic synthetic names to every named
-/// feature in the bundle. Seeded RNG picks each name from a small
-/// per-category pool so a given `seed` always produces the same name
-/// set. Settlement IDs disambiguate when the pool is smaller than the
-/// settlement count (e.g. `"Brightford-7"`).
-///
-/// **NOT an LLM call.** LLM-driven naming requires a new `TextProvider`
-/// trait (the v4.3 [`crate::shape::llm::LlmProvider`] is structured for
-/// ShapeKind picking, not free-form text generation). Ship 7b will add
-/// that trait + Anthropic / OpenAI / Ollama impls and replace this
-/// stub. Until then synthetic names keep SVG export (Ship 9) labelable.
-#[allow(clippy::too_many_arguments)]
-pub fn apply_synthetic_names(
-    features: &mut Features,
-    political: &mut Political,
-    settlements: &mut [Settlement],
-    culture: &mut Culture,
-    seed: u64,
-) {
-    let mut rng = Rng::for_stage(seed, b"civ-naming");
-    for (i, s) in settlements.iter_mut().enumerate() {
-        s.name = format!("{}-{i}", pick(SETTLEMENT_NAMES, &mut rng));
-    }
-    for (i, st) in political.states.iter_mut().enumerate() {
-        st.name = format!("{}-{i}", pick(STATE_NAMES, &mut rng));
-    }
-    for (i, p) in political.provinces.iter_mut().enumerate() {
-        let prefix = pick(PROVINCE_PREFIXES, &mut rng);
-        let root = pick(PROVINCE_ROOTS, &mut rng);
-        p.name = format!("{prefix} {root}-{i}");
-    }
-    for (i, c) in culture.culture_regions.iter_mut().enumerate() {
-        c.name = format!("{}-{i}", pick(CULTURE_NAMES, &mut rng));
-    }
-    for (i, mr) in features.mountain_ranges.iter_mut().enumerate() {
-        mr.name = format!(
-            "{} Mountains-{i}",
-            pick(MOUNTAIN_DESCRIPTORS, &mut rng)
-        );
-    }
-    for (i, rv) in features.rivers.iter_mut().enumerate() {
-        rv.name = format!("{} River-{i}", pick(RIVER_DESCRIPTORS, &mut rng));
-    }
-    for (i, wb) in features.water_bodies.iter_mut().enumerate() {
-        wb.name = format!("{}-{i}", pick(WATER_BODY_DESCRIPTORS, &mut rng));
-    }
-}
-
-/// **Civ Ship 5** — full pipeline through System-A's routes builder.
-/// Chains [`build_settlement`] then [`routes::build`] using the
-/// `Hydrology.river_threshold` for river-route detection.
-pub fn build_routes(
-    world: &FlatWorld,
-    climate_params: &WorldClimateParams,
-    ocean_target: usize,
-    seed: u64,
-    density: SettlementDensity,
-) -> (
-    CivView,
-    Features,
-    Political,
-    Hydrology,
-    Vec<Settlement>,
-    Vec<Route>,
-) {
-    let (view, features, political, hydro, settlements) =
-        build_settlement(world, climate_params, ocean_target, seed, density);
-    let routes_v = routes::build(
-        &view.centers,
-        &view.neighbors,
-        &view.biomes,
-        &view.river_flux,
-        hydro.river_threshold,
-        &view.is_coast,
-        &settlements,
-    );
-    (view, features, political, hydro, settlements, routes_v)
-}
-
-/// **Civ Ship 6** — full pipeline through System-A's culture builder.
-/// Adds [`culture::build`] on top of [`build_routes`].
-///
-/// `culture_count` is the requested number of distinct cultures
-/// (clamped 1..=16 internally by System A). 5 is the typical default.
-#[allow(clippy::too_many_arguments)]
-pub fn build_culture(
-    world: &FlatWorld,
-    climate_params: &WorldClimateParams,
-    ocean_target: usize,
-    seed: u64,
-    density: SettlementDensity,
-    culture_count: u8,
-) -> (
-    CivView,
-    Features,
-    Political,
-    Hydrology,
-    Vec<Settlement>,
-    Vec<Route>,
-    Culture,
-) {
-    let (view, features, political, hydro, settlements, routes_v) =
-        build_routes(world, climate_params, ocean_target, seed, density);
-    let culture_v = culture::build(seed, &view.centers, &view.neighbors, &view.biomes, culture_count);
-    (
-        view,
-        features,
-        political,
-        hydro,
-        settlements,
-        routes_v,
-        culture_v,
-    )
-}
-
-/// **Civ Ship 3** — full pipeline through System-A's political builder.
-/// Computes the CivView, augments with synthetic ocean cells, extracts
-/// features, then derives provinces + states via [`political::build`].
-///
-/// Returns `(view, features, political)` so callers can pass any
-/// combination into the downstream settlement / route / culture
-/// builders.
-pub fn build_political(
-    world: &FlatWorld,
-    climate_params: &WorldClimateParams,
-    ocean_target: usize,
-    seed: u64,
-) -> (CivView, Features, Political) {
-    let (view, features) = extract_features(world, climate_params, ocean_target);
-    let political = political::build(seed, &view.centers, &view.neighbors, &view.biomes);
-    (view, features, political)
-}
-
-/// Build neighbor lists from a Delaunay triangulation over the 2D cell
-/// centres. The triangulation's edge set becomes the adjacency.
-fn build_delaunay_neighbors(centers_2d: &[(f32, f32)]) -> Vec<Vec<u32>> {
-    let n = centers_2d.len();
-    let mut neighbors: Vec<Vec<u32>> = vec![Vec::new(); n];
-    if n < 3 {
-        // Delaunay needs at least 3 points; degenerate cases produce no
-        // edges. The civilization stack still works on a disconnected
-        // graph — it just won't have any provinces / settlements.
-        return neighbors;
-    }
-    let points: Vec<Point> = centers_2d
-        .iter()
-        .map(|&(x, y)| Point { x: x as f64, y: y as f64 })
-        .collect();
-    let tri = triangulate(&points);
-    // Each triangle is 3 consecutive entries in `tri.triangles`.
-    for t in tri.triangles.chunks_exact(3) {
-        let (a, b, c) = (t[0] as u32, t[1] as u32, t[2] as u32);
-        for (u, v) in [(a, b), (b, c), (c, a)] {
-            neighbors[u as usize].push(v);
-            neighbors[v as usize].push(u);
-        }
-    }
-    for nb in neighbors.iter_mut() {
-        nb.sort_unstable();
-        nb.dedup();
-    }
-    neighbors
-}
-
-/// Cells that sit next to an ocean cell are coast — `settlement::build`
-/// biases toward these.
-fn compute_is_coast(biomes: &[BiomeKind], neighbors: &[Vec<u32>]) -> Vec<bool> {
-    let mut out = vec![false; biomes.len()];
-    for (i, nb) in neighbors.iter().enumerate() {
-        if biomes[i] == BiomeKind::Ocean {
-            continue; // ocean itself isn't "coast"
-        }
-        if nb
-            .iter()
-            .any(|&j| matches!(biomes[j as usize], BiomeKind::Ocean))
-        {
-            out[i] = true;
-        }
-    }
-    out
-}
-
-/// Lossy mapping from the flatworld 21-cell Köppen classification to the
-/// System-A 14-cell coarse biome. Trade-off: political / settlement /
-/// route algorithms care about Forest-vs-Plain-vs-Desert distinctions,
-/// not the Köppen subtypes that drive biome render colours.
-pub fn koppen_to_biome_kind(koppen: Biome) -> BiomeKind {
-    match koppen {
-        // Polar
-        Biome::Ef => BiomeKind::Glacier,
-        Biome::Et => BiomeKind::Tundra,
-        // Continental
-        Biome::Dfd | Biome::Dfc => BiomeKind::Forest, // boreal forest (taiga)
-        Biome::Dfb | Biome::Dfa | Biome::Dwa => BiomeKind::Forest,
-        // Temperate
-        Biome::Cfb | Biome::Cfa | Biome::Cwa => BiomeKind::Forest,
-        Biome::Csa | Biome::Csb => BiomeKind::Plain, // Mediterranean — open
-        // Arid
-        Biome::Bsh | Biome::Bsk => BiomeKind::Plain, // steppe → grassland
-        Biome::Bwh | Biome::Bwk => BiomeKind::Desert,
-        // Tropical
-        Biome::Af | Biome::Am => BiomeKind::Jungle,
-        Biome::Aw => BiomeKind::Plain, // savanna
-    }
-}
-
-/// Coarse System-A `ClimateZone` from per-cell temperature + precipitation
-/// — the simple bucket-classifier System A's `climate::build` would have
-/// produced with the same inputs.
-pub fn derive_climate_zone(temp_mean: f32, precip_annual: f32) -> ClimateZone {
-    if temp_mean < -5.0 {
-        return ClimateZone::Polar;
-    }
-    if temp_mean < 5.0 {
-        return ClimateZone::Boreal;
-    }
-    if precip_annual < 250.0 {
-        return ClimateZone::Arid;
-    }
-    if temp_mean > 22.0 {
-        return ClimateZone::Tropical;
-    }
-    if temp_mean > 15.0 {
-        if precip_annual < 600.0 {
-            ClimateZone::Mediterranean
-        } else {
-            ClimateZone::Subtropical
-        }
-    } else {
-        ClimateZone::Temperate
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::feature;
-    use crate::flatworld::{generate, FlatParams};
-
-    fn small_world() -> FlatWorld {
-        // Small but multi-plate so neighbors / coast detection have
-        // something to chew on — 4 plates × default zone/sub counts is
-        // ~30-40 cells.
-        let p = FlatParams {
-            width: 256,
-            height: 192,
-            plate_count: 4,
-            seed: 7,
-            ..Default::default()
-        };
-        generate(&p)
-    }
-
-    #[test]
-    fn civ_view_has_one_cell_per_subzone() {
-        let world = small_world();
-        let view = build_civ_view(&world, &WorldClimateParams::default());
-        let expected: usize = world
-            .plates
-            .iter()
-            .flat_map(|p| p.zones.iter())
-            .map(|z| z.subzones.len())
-            .sum();
-        assert_eq!(view.centers.len(), expected);
-        assert_eq!(view.neighbors.len(), expected);
-        assert_eq!(view.biomes.len(), expected);
-        assert_eq!(view.climate.len(), expected);
-        assert_eq!(view.river_flux.len(), expected);
-        assert_eq!(view.is_coast.len(), expected);
-        assert_eq!(view.elevation.len(), expected);
-    }
-
-    #[test]
-    fn delaunay_neighbors_are_symmetric_and_deduped() {
-        let world = small_world();
-        let view = build_civ_view(&world, &WorldClimateParams::default());
-        for (i, nb) in view.neighbors.iter().enumerate() {
-            let mut sorted = nb.clone();
-            sorted.sort_unstable();
-            assert_eq!(*nb, sorted, "cell {i}: neighbors should be sorted");
-            assert!(
-                nb.windows(2).all(|w| w[0] != w[1]),
-                "cell {i}: neighbors should be deduped"
-            );
-            for &j in nb {
-                assert!(
-                    view.neighbors[j as usize].contains(&(i as u32)),
-                    "adjacency asymmetric: {i} -> {j} but not back"
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn delaunay_produces_planar_density() {
-        // A Delaunay triangulation over interior points has on average ≈6
-        // neighbors per vertex (Euler's formula, planar graph). Boundary
-        // vertices have fewer. The average across the view should land in
-        // a sane window — proves the triangulation is actually building
-        // edges, not returning an empty mesh.
-        let world = small_world();
-        let view = build_civ_view(&world, &WorldClimateParams::default());
-        if view.centers.len() < 4 {
-            return; // degenerate small test world
-        }
-        let total_edges: usize = view.neighbors.iter().map(|nb| nb.len()).sum();
-        let avg = total_edges as f32 / view.centers.len() as f32;
-        assert!(
-            (3.0..=12.0).contains(&avg),
-            "average neighbors per cell {avg:.2} outside [3, 12]; planar graph density is broken",
-        );
-    }
-
-    #[test]
-    fn feature_extract_accepts_civ_view_without_panicking() {
-        // Smoke: System-A's mesh-agnostic feature::extract should accept
-        // the adapter's (biomes, neighbors) and not panic. We don't pin
-        // a specific feature count here — Ship 1's civilization adapter
-        // only emits BiomeKind from L2 sub-zone centres, and those are
-        // by construction inside plates (no synthetic Ocean cells yet),
-        // so feature counts can be zero on small test worlds. Synthetic
-        // ocean/river adjacency lands in civ-layer Ship 2.
-        let world = small_world();
-        let view = build_civ_view(&world, &WorldClimateParams::default());
-        let features = feature::extract(&view.biomes, &view.neighbors);
-        // Just touch the result so the compiler doesn't optimise it out.
-        let _ = features.mountain_ranges.len()
-            + features.rivers.len()
-            + features.water_bodies.len();
-    }
-
-    #[test]
-    fn koppen_translation_covers_every_input_variant() {
-        // Exhaustive: every Köppen variant must map to *some* BiomeKind.
-        // Compiler enforcement — adding a Köppen variant fails this test
-        // with E0004 until the match arm is added in
-        // koppen_to_biome_kind.
-        let variants = [
-            Biome::Ef, Biome::Et,
-            Biome::Dfd, Biome::Dfc, Biome::Dfb, Biome::Dfa, Biome::Dwa,
-            Biome::Cfb, Biome::Cfa, Biome::Csa, Biome::Csb, Biome::Cwa,
-            Biome::Bsk, Biome::Bwk, Biome::Bsh, Biome::Bwh,
-            Biome::Af, Biome::Am, Biome::Aw,
-        ];
-        for v in variants {
-            let _ = koppen_to_biome_kind(v);
-        }
-    }
-
-    #[test]
-    fn climate_zone_derivation_hits_each_bucket() {
-        // Smoke that the cutoffs aren't tangled: a hot wet point lands in
-        // Tropical; a cold point lands in Polar; etc.
-        assert_eq!(derive_climate_zone(-20.0, 200.0), ClimateZone::Polar);
-        assert_eq!(derive_climate_zone(0.0, 600.0), ClimateZone::Boreal);
-        assert_eq!(derive_climate_zone(10.0, 100.0), ClimateZone::Arid);
-        assert_eq!(derive_climate_zone(28.0, 2000.0), ClimateZone::Tropical);
-        assert_eq!(derive_climate_zone(18.0, 400.0), ClimateZone::Mediterranean);
-        assert_eq!(derive_climate_zone(20.0, 1200.0), ClimateZone::Subtropical);
-        assert_eq!(derive_climate_zone(12.0, 1200.0), ClimateZone::Temperate);
-    }
-
-    #[test]
-    fn coast_flag_set_for_cells_neighbouring_ocean() {
-        // If any cell is ocean and any of its neighbours is land, the
-        // land neighbour must be flagged is_coast. If no ocean cell
-        // exists in the test world the assertion vacuously holds.
-        let world = small_world();
-        let view = build_civ_view(&world, &WorldClimateParams::default());
-        for (i, &biome) in view.biomes.iter().enumerate() {
-            if biome == BiomeKind::Ocean {
-                for &nb in &view.neighbors[i] {
-                    let j = nb as usize;
-                    if view.biomes[j] != BiomeKind::Ocean {
-                        assert!(
-                            view.is_coast[j],
-                            "cell {j} (biome {:?}) is adjacent to ocean cell {i} \
-                             but is_coast is false",
-                            view.biomes[j]
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn augment_with_ocean_adds_synthetic_ocean_cells() {
-        // Ship 2: augment pass should append BiomeKind::Ocean cells AND
-        // those cells should fall in void area (no plate covers them).
-        let world = small_world();
-        let view = build_civ_view(&world, &WorldClimateParams::default());
-        let land_n = view.centers.len();
-        let augmented = augment_with_ocean(view, &world, 16);
-        let total_n = augmented.centers.len();
-        assert!(
-            total_n > land_n,
-            "augment should grow the cell count; before {land_n}, after {total_n}",
-        );
-        let ocean_n = augmented
-            .biomes
-            .iter()
-            .filter(|&&b| b == BiomeKind::Ocean)
-            .count();
-        assert!(ocean_n >= 1, "should have ≥1 synthetic Ocean cell");
-        // Each synthetic ocean cell's centre must be in void.
-        for (i, &biome) in augmented.biomes.iter().enumerate().skip(land_n) {
-            assert_eq!(biome, BiomeKind::Ocean, "appended cell {i} should be Ocean");
-            let [x, y, _] = augmented.centers[i];
-            assert!(
-                world.plates_at(x, y).is_empty(),
-                "synthetic ocean cell {i} at ({x:.1},{y:.1}) lands inside a plate"
-            );
-        }
-    }
-
-    #[test]
-    fn augment_with_ocean_flips_coast_flag_on_neighbouring_land() {
-        // After ocean augmentation a coastal land sub-zone (adjacent to
-        // at least one synthetic ocean cell) must have is_coast=true.
-        let world = small_world();
-        let view = build_civ_view(&world, &WorldClimateParams::default());
-        let land_n = view.centers.len();
-        let augmented = augment_with_ocean(view, &world, 16);
-        let mut coast_hits = 0usize;
-        for (i, &biome) in augmented.biomes.iter().enumerate().take(land_n) {
-            if biome == BiomeKind::Ocean {
-                continue;
-            }
-            let touches_ocean = augmented.neighbors[i]
-                .iter()
-                .any(|&j| augmented.biomes[j as usize] == BiomeKind::Ocean);
-            if touches_ocean {
-                assert!(
-                    augmented.is_coast[i],
-                    "land cell {i} touches ocean but is_coast is false"
-                );
-                coast_hits += 1;
-            }
-        }
-        assert!(
-            coast_hits > 0,
-            "expected at least one coast cell after augment; got 0"
-        );
-    }
-
-    #[test]
-    fn augment_with_ocean_is_deterministic_per_target() {
-        // Same world + same target_count must produce the same ocean
-        // point set (Civ-layer must be byte-deterministic to keep the
-        // generate→export pipeline reproducible).
-        let world = small_world();
-        let a = augment_with_ocean(
-            build_civ_view(&world, &WorldClimateParams::default()),
-            &world,
-            12,
-        );
-        let b = augment_with_ocean(
-            build_civ_view(&world, &WorldClimateParams::default()),
-            &world,
-            12,
-        );
-        assert_eq!(a.centers, b.centers);
-        assert_eq!(a.biomes, b.biomes);
-    }
-
-    #[test]
-    fn extract_features_default_world_yields_water_and_mountains() {
-        // Ship 2 acceptance: on a default-sized world the convenience
-        // pipeline must produce ≥1 water_body AND ≥1 mountain_range —
-        // proves the (ocean synthesis + biome derivation + Delaunay
-        // adjacency + feature::extract) chain actually does work
-        // end-to-end. Smaller worlds were too sparse in Ship 1.
-        let world = generate(&FlatParams::default());
-        let (_view, features) = extract_features(&world, &WorldClimateParams::default(), 64);
-        assert!(
-            !features.water_bodies.is_empty(),
-            "default world should produce ≥1 water body after ocean synth"
-        );
-        assert!(
-            !features.mountain_ranges.is_empty(),
-            "default world should produce ≥1 mountain range from collision uplift"
-        );
-    }
-
-    #[test]
-    fn extract_features_with_zero_ocean_target_skips_augment() {
-        // ocean_target=0 should be a no-op (matches Ship 1 behaviour).
-        let world = small_world();
-        let view_ship1 = build_civ_view(&world, &WorldClimateParams::default());
-        let (view_skip, _) = extract_features(&world, &WorldClimateParams::default(), 0);
-        assert_eq!(view_skip.centers.len(), view_ship1.centers.len());
-        assert_eq!(view_skip.biomes, view_ship1.biomes);
-    }
-
-    #[test]
-    fn build_political_produces_provinces_and_states_on_default_world() {
-        // Ship 3 acceptance: System-A's political::build, fed the civ
-        // adapter output, must produce ≥1 province AND ≥1 state on a
-        // default-sized world. Small worlds may degenerate (too few land
-        // cells for the province/state thresholds) so this MUST run
-        // against the default `FlatParams`.
-        let world = generate(&FlatParams::default());
-        let (_view, _features, political) = build_political(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-        );
-        assert!(
-            !political.provinces.is_empty(),
-            "default world should produce ≥1 province"
-        );
-        assert!(
-            !political.states.is_empty(),
-            "default world should produce ≥1 state"
-        );
-    }
-
-    #[test]
-    fn build_political_assigns_every_land_cell_to_a_province() {
-        // Every land cell must have province_of != NONE (NONE = u32::MAX
-        // by convention in world_map.rs). Ocean cells stay NONE — they
-        // belong to no province.
-        let world = generate(&FlatParams::default());
-        let (view, _features, political) = build_political(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            7,
-        );
-        let none = u32::MAX;
-        for (i, &biome) in view.biomes.iter().enumerate() {
-            let p = political.province_of[i];
-            if biome == BiomeKind::Ocean {
-                // Ocean cells should be unassigned.
-                assert_eq!(
-                    p, none,
-                    "ocean cell {i} got assigned province {p}",
-                );
-            } else {
-                assert_ne!(
-                    p, none,
-                    "land cell {i} (biome {biome:?}) has no province assignment",
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn build_political_is_deterministic_per_seed() {
-        // Same (world, climate, ocean_target, seed) → same political
-        // output. Civ-layer pipeline must stay byte-deterministic to
-        // preserve the generate→export reproducibility guarantee.
-        let world = generate(&FlatParams::default());
-        let (_, _, a) = build_political(&world, &WorldClimateParams::default(), 32, 99);
-        let (_, _, b) = build_political(&world, &WorldClimateParams::default(), 32, 99);
-        assert_eq!(a.province_of, b.province_of);
-        assert_eq!(a.provinces.len(), b.provinces.len());
-        assert_eq!(a.states.len(), b.states.len());
-    }
-
-    #[test]
-    fn elevation_to_u16_keeps_void_below_sea_and_base_above() {
-        // Void/synthetic-ocean cells (elev 0.0) must end up STRICTLY
-        // below sea_level_u16; cells at BASE_LEVEL (0.35) must end up
-        // STRICTLY above. This is the contract Ship 4 relies on so
-        // hydrology's priority-flood correctly identifies ocean cells.
-        let world = generate(&FlatParams::default());
-        let view = augment_with_ocean(
-            build_civ_view(&world, &WorldClimateParams::default()),
-            &world,
-            32,
-        );
-        let (elev_u16, sea) = elevation_to_u16(&view);
-        for (i, &biome) in view.biomes.iter().enumerate() {
-            if biome == BiomeKind::Ocean {
-                assert!(
-                    elev_u16[i] < sea,
-                    "ocean cell {i} elev_u16 {} should be < sea {sea}",
-                    elev_u16[i]
-                );
-            } else {
-                assert!(
-                    elev_u16[i] >= sea,
-                    "land cell {i} (biome {biome:?}) elev_u16 {} should be ≥ sea {sea}",
-                    elev_u16[i]
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn build_settlement_produces_settlements_on_default_world() {
-        // Ship 4 acceptance: System-A's settlement::build, fed the civ
-        // adapter output + hydrology-derived river_flux + is_coast,
-        // must produce ≥3 settlements on a default-sized world.
-        let world = generate(&FlatParams::default());
-        let (_view, _f, _pol, _hyd, settlements) = build_settlement(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-            SettlementDensity::Medium,
-        );
-        assert!(
-            settlements.len() >= 3,
-            "expected ≥3 settlements on default world, got {}",
-            settlements.len()
-        );
-    }
-
-    #[test]
-    fn settlements_only_land_cells_not_ocean() {
-        let world = generate(&FlatParams::default());
-        let (view, _, _, _, settlements) = build_settlement(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            7,
-            SettlementDensity::Medium,
-        );
-        for s in &settlements {
-            let cell = s.cell as usize;
-            assert_ne!(
-                view.biomes[cell],
-                BiomeKind::Ocean,
-                "settlement '{}' placed on Ocean cell {cell}",
-                s.name
-            );
-        }
-    }
-
-    #[test]
-    fn build_settlement_is_deterministic_per_seed() {
-        let world = generate(&FlatParams::default());
-        let (_, _, _, _, a) = build_settlement(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            99,
-            SettlementDensity::Medium,
-        );
-        let (_, _, _, _, b) = build_settlement(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            99,
-            SettlementDensity::Medium,
-        );
-        assert_eq!(a.len(), b.len());
-        for (sa, sb) in a.iter().zip(b.iter()) {
-            assert_eq!(sa.cell, sb.cell);
-            assert_eq!(sa.role, sb.role);
-        }
-    }
-
-    #[test]
-    fn build_routes_produces_routes_on_default_world() {
-        // Ship 5 acceptance: System-A's routes::build, fed the full civ
-        // pipeline, must produce ≥1 Route on a default-sized world.
-        let world = generate(&FlatParams::default());
-        let (_, _, _, _, _, routes_v) = build_routes(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-            SettlementDensity::Medium,
-        );
-        assert!(
-            !routes_v.is_empty(),
-            "default world should produce ≥1 route"
-        );
-    }
-
-    #[test]
-    fn build_culture_produces_multiple_regions_on_default_world() {
-        // Ship 6 acceptance: System-A's culture::build, fed the full civ
-        // pipeline, must produce ≥2 culture regions when culture_count=5
-        // is requested on a default-sized world.
-        let world = generate(&FlatParams::default());
-        let (_, _, _, _, _, _, culture_v) = build_culture(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-            SettlementDensity::Medium,
-            5,
-        );
-        assert!(
-            culture_v.culture_regions.len() >= 2,
-            "default world should produce ≥2 culture regions, got {}",
-            culture_v.culture_regions.len()
-        );
-    }
-
-    #[test]
-    fn build_culture_is_deterministic_per_seed() {
-        let world = generate(&FlatParams::default());
-        let (_, _, _, _, _, _, a) = build_culture(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            99,
-            SettlementDensity::Medium,
-            5,
-        );
-        let (_, _, _, _, _, _, b) = build_culture(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            99,
-            SettlementDensity::Medium,
-            5,
-        );
-        assert_eq!(a.culture_of, b.culture_of);
-        assert_eq!(a.culture_regions.len(), b.culture_regions.len());
-    }
-
-    #[test]
-    fn synthetic_names_populate_every_named_feature() {
-        // Ship 7: every named feature in the bundle must end up with a
-        // non-empty `name` after `apply_synthetic_names`. Empty names
-        // would crash the SVG export label rendering in Ship 9.
-        let world = generate(&FlatParams::default());
-        let (_view, mut features, mut political, _hyd, mut settlements, _routes_v, mut culture_v) =
-            build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                64,
-                42,
-                SettlementDensity::Medium,
-                5,
-            );
-        apply_synthetic_names(
-            &mut features,
-            &mut political,
-            &mut settlements,
-            &mut culture_v,
-            42,
-        );
-        for s in &settlements {
-            assert!(!s.name.is_empty(), "settlement {} unnamed", s.cell);
-        }
-        for st in &political.states {
-            assert!(!st.name.is_empty(), "state {} unnamed", st.id);
-        }
-        for p in &political.provinces {
-            assert!(!p.name.is_empty(), "province {} unnamed", p.id);
-        }
-        for c in &culture_v.culture_regions {
-            assert!(!c.name.is_empty(), "culture {} unnamed", c.id);
-        }
-        for mr in &features.mountain_ranges {
-            assert!(!mr.name.is_empty(), "mountain {} unnamed", mr.id);
-        }
-        for wb in &features.water_bodies {
-            assert!(!wb.name.is_empty(), "water body {} unnamed", wb.id);
-        }
-    }
-
-    #[test]
-    fn synthetic_names_are_deterministic_per_seed() {
-        // Same seed → same names.
-        let world = generate(&FlatParams::default());
-
-        let make_bundle = || {
-            let (_, f, p, _, s, _, c) = build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                32,
-                99,
-                SettlementDensity::Medium,
-                5,
-            );
-            (f, p, s, c)
-        };
-        let (mut fa, mut pa, mut sa, mut ca) = make_bundle();
-        let (mut fb, mut pb, mut sb, mut cb) = make_bundle();
-        apply_synthetic_names(&mut fa, &mut pa, &mut sa, &mut ca, 99);
-        apply_synthetic_names(&mut fb, &mut pb, &mut sb, &mut cb, 99);
-        for (a, b) in sa.iter().zip(sb.iter()) {
-            assert_eq!(a.name, b.name);
-        }
-        for (a, b) in pa.provinces.iter().zip(pb.provinces.iter()) {
-            assert_eq!(a.name, b.name);
-        }
-    }
-
-    #[test]
-    fn name_civ_via_llm_with_mock_populates_every_category() {
-        // Ship 7b: name_civ_via_llm with MockTextProvider must fill in
-        // names for every named feature. Same property as Ship 7
-        // synthetic naming, just via the trait path.
-        use crate::shape::llm::MockTextProvider;
-        let world = generate(&FlatParams::default());
-        let (_, mut features, mut political, _, mut settlements, _routes_v, mut culture_v) =
-            build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                64,
-                42,
-                SettlementDensity::Medium,
-                5,
-            );
-        let provider = MockTextProvider::new();
-        name_civ_via_llm(
-            &mut features,
-            &mut political,
-            &mut settlements,
-            &mut culture_v,
-            &provider,
-            "HighFantasy",
-        )
-        .expect("mock should succeed");
-
-        // Mock returns exactly 5 entries per category — every feature
-        // up to index 5 should be named; surplus features stay unnamed.
-        let named_settlements = settlements.iter().filter(|s| !s.name.is_empty()).count();
-        let named_states = political.states.iter().filter(|s| !s.name.is_empty()).count();
-        assert!(
-            named_settlements >= 1,
-            "expected ≥1 named settlement after mock LLM naming"
-        );
-        assert!(
-            named_states >= 1,
-            "expected ≥1 named state after mock LLM naming"
-        );
-    }
-
-    #[test]
-    fn name_civ_via_llm_is_deterministic_with_mock_and_archetype() {
-        use crate::shape::llm::MockTextProvider;
-        let world = generate(&FlatParams::default());
-
-        let make_bundle = || {
-            let (_, f, p, _, s, _, c) = build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                32,
-                99,
-                SettlementDensity::Medium,
-                5,
-            );
-            (f, p, s, c)
-        };
-        let (mut fa, mut pa, mut sa, mut ca) = make_bundle();
-        let (mut fb, mut pb, mut sb, mut cb) = make_bundle();
-        let provider = MockTextProvider::new();
-        name_civ_via_llm(&mut fa, &mut pa, &mut sa, &mut ca, &provider, "HighFantasy").unwrap();
-        name_civ_via_llm(&mut fb, &mut pb, &mut sb, &mut cb, &provider, "HighFantasy").unwrap();
-        for (a, b) in sa.iter().zip(sb.iter()) {
-            assert_eq!(a.name, b.name, "settlement name should match across runs");
-        }
-        for (a, b) in pa.provinces.iter().zip(pb.provinces.iter()) {
-            assert_eq!(a.name, b.name);
-        }
-    }
-
-    #[test]
-    fn name_civ_via_llm_differs_across_archetypes_via_mock_hash() {
-        // Different archetype string → different user prompt → different
-        // mock hash → different name set. Proves the archetype actually
-        // flows into the provider call.
-        use crate::shape::llm::MockTextProvider;
-        let world = generate(&FlatParams::default());
-
-        let make_bundle = || {
-            let (_, f, p, _, s, _, c) = build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                32,
-                7,
-                SettlementDensity::Medium,
-                5,
-            );
-            (f, p, s, c)
-        };
-        let (mut fa, mut pa, mut sa, mut ca) = make_bundle();
-        let (mut fb, mut pb, mut sb, mut cb) = make_bundle();
-        let provider = MockTextProvider::new();
-        name_civ_via_llm(&mut fa, &mut pa, &mut sa, &mut ca, &provider, "HighFantasy").unwrap();
-        name_civ_via_llm(&mut fb, &mut pb, &mut sb, &mut cb, &provider, "Cyberpunk").unwrap();
-        let differ = sa.iter().zip(sb.iter()).any(|(a, b)| a.name != b.name);
-        assert!(
-            differ,
-            "two distinct archetypes produced identical settlement names; \
-             archetype not reaching prompt"
-        );
-    }
-
-    #[test]
-    fn name_civ_via_llm_returns_invalid_response_on_garbage() {
-        // A custom provider that returns junk must surface as
-        // LlmError::InvalidResponse — caller can then fall back to
-        // synthetic naming.
-        use crate::shape::llm::{LlmError, TextPrompt, TextProvider};
-
-        #[derive(Debug)]
-        struct GarbageProvider;
-        impl TextProvider for GarbageProvider {
-            fn complete(&self, _: &TextPrompt) -> Result<String, LlmError> {
-                Ok("definitely not json {".to_string())
-            }
-        }
-
-        let world = generate(&FlatParams::default());
-        let (_, mut features, mut political, _, mut settlements, _, mut culture_v) =
-            build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                32,
-                7,
-                SettlementDensity::Medium,
-                5,
-            );
-        let err = name_civ_via_llm(
-            &mut features,
-            &mut political,
-            &mut settlements,
-            &mut culture_v,
-            &GarbageProvider,
-            "HighFantasy",
-        )
-        .expect_err("garbage should fail");
-        assert!(matches!(err, LlmError::InvalidResponse(_)));
-    }
-
-    #[test]
-    fn civ_bundle_json_round_trip_preserves_content_hash() {
-        // Ship 9 acceptance: serialize CivBundle to JSON → deserialize
-        // back → verify_civ_hash holds. The round-trip must be
-        // byte-stable for the content_hash to be load-bearing.
-        let world = generate(&FlatParams::default());
-        let bundle = bundle_civ(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-            SettlementDensity::Medium,
-            5,
-        );
-        assert!(verify_civ_hash(&bundle), "fresh bundle should verify");
-        let json = serde_json::to_string(&bundle).expect("serialize");
-        let parsed: CivBundle = serde_json::from_str(&json).expect("deserialize");
-        assert_eq!(
-            parsed.content_hash, bundle.content_hash,
-            "round-trip content_hash should be identical"
-        );
-        assert!(
-            verify_civ_hash(&parsed),
-            "deserialized bundle should still verify against the hash"
-        );
-        assert_eq!(parsed, bundle, "round-trip bundle equality");
-    }
-
-    #[test]
-    fn civ_bundle_hash_is_deterministic_per_seed() {
-        let world = generate(&FlatParams::default());
-        let a = bundle_civ(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            99,
-            SettlementDensity::Medium,
-            5,
-        );
-        let b = bundle_civ(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            99,
-            SettlementDensity::Medium,
-            5,
-        );
-        assert_eq!(a.content_hash, b.content_hash);
-    }
-
-    #[test]
-    fn spherical_centers_lie_on_unit_sphere() {
-        // Approach A step 2: every projected centre must have |c|² ≈ 1
-        // within a small float-rounding window. Proves the equirectangular
-        // → sphere math is correct end-to-end.
-        let world = generate(&FlatParams::default());
-        let view = build_civ_view_spherical(&world, &WorldClimateParams::default());
-        for (i, c) in view.centers.iter().enumerate() {
-            let mag2 = c[0] * c[0] + c[1] * c[1] + c[2] * c[2];
-            assert!(
-                (mag2 - 1.0).abs() < 1e-4,
-                "cell {i} centre {:?} has |c|² = {mag2}; should be 1.0",
-                c
-            );
-        }
-    }
-
-    #[test]
-    fn spherical_centers_preserve_distinctness() {
-        // The flat centres are non-degenerate (no two subzones share a
-        // centre). After equirectangular projection they must STAY
-        // distinct enough for distance-based algorithms (political
-        // farthest-point sampling, settlement-capital matching) to
-        // distinguish them.
-        //
-        // **review-impl MED-2 fix**: previous test used 1µ
-        // quantisation (1e-6). On default 1280×720 world the
-        // equirectangular per-pixel pitch is ~5 mrad → ~5000µ on the
-        // unit sphere, so 1µ would only catch literal identical
-        // [f32; 3] collisions. Tightened to 1 mrad quantisation so the
-        // test fails if any two centres land within ~1 mrad of each
-        // other — meaningfully smaller than the per-pixel pitch and
-        // well above f32 round-off.
-        let world = generate(&FlatParams::default());
-        let view = build_civ_view_spherical(&world, &WorldClimateParams::default());
-        let mut seen: std::collections::HashSet<(i32, i32, i32)> =
-            std::collections::HashSet::new();
-        let q = 1_000.0_f32; // 1 mrad quantisation; arc length 1 µ on a unit sphere
-        for c in &view.centers {
-            let key = (
-                (c[0] * q) as i32,
-                (c[1] * q) as i32,
-                (c[2] * q) as i32,
-            );
-            assert!(
-                seen.insert(key),
-                "two sphere centres landed within 1 mrad of each other \
-                 at quantised key {:?}; equirectangular projection has a \
-                 near-degenerate collision",
-                key
-            );
-        }
-    }
-
-    #[test]
-    fn bundle_civ_emits_sphere_centers_post_review_fix() {
-        // **review-impl HIGH-1 regression test (1 of 2)**. After the
-        // fix, `extract_features` (and therefore every convenience
-        // function chained off it — bundle_civ / build_political /
-        // build_settlement / build_routes / build_culture) project to
-        // sphere before downstream civilization builders run, so the
-        // unit-vector distance metric they use is meaningful.
-        //
-        // Direct evidence of the fix: every `bundle.centers[i]` is a
-        // unit vector (|c|² ≈ 1). Pre-fix it was `[x, y, 0]` with
-        // |c|² ≈ x² + y² ranging into the millions.
-        let world = generate(&FlatParams::default());
-        let bundle = bundle_civ(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-            SettlementDensity::Medium,
-            5,
-        );
-        for (i, c) in bundle.centers.iter().enumerate() {
-            let mag2 = c[0] * c[0] + c[1] * c[1] + c[2] * c[2];
-            assert!(
-                (mag2 - 1.0).abs() < 1e-4,
-                "bundle.centers[{i}] = {:?} has |c|² = {mag2}; expected unit vector \
-                 (review-impl HIGH-1 fix not applied)",
-                c
-            );
-        }
-    }
-
-    #[test]
-    fn sphere_default_does_not_worsen_political_seed_spread_vs_flat() {
-        // **review-impl HIGH-1 regression test (2 of 2)**. The metric
-        // mismatch on flat substrate is real but small in impact —
-        // `political::build`'s pass-2 random-fill fallback masks most
-        // of the degenerate spaced_ok behaviour, so empirically sphere
-        // only nudges the spread ~3-5% on default worlds. This test
-        // pins the *direction* of the fix (sphere ≥ flat for the
-        // min-pairwise distance between province capital cells) so
-        // future refactors can't silently regress the metric back to
-        // a flat-substrate path.
-        //
-        // World pixel coordinates of capital cells are stable across
-        // substrates (cell_index → subzone.center is unchanged), so
-        // we can compare both branches via the same metric.
-        let world = generate(&FlatParams::default());
-        let climate = WorldClimateParams::default();
-
-        let flat_view = augment_with_ocean(
-            build_civ_view(&world, &climate),
-            &world,
-            64,
-        );
-        let flat_pol = crate::political::build(
-            42,
-            &flat_view.centers,
-            &flat_view.neighbors,
-            &flat_view.biomes,
-        );
-
-        let (_, _, sphere_pol) = build_political(&world, &climate, 64, 42);
-
-        if flat_pol.provinces.len() < 2 || sphere_pol.provinces.len() < 2 {
-            return; // degenerate; nothing to compare.
-        }
-
-        let capital_xy = |political: &crate::political::Political| {
-            political
-                .provinces
-                .iter()
-                .filter_map(|p| cell_index_to_center(&world, p.capital_cell as usize))
-                .collect::<Vec<_>>()
-        };
-        let min_pairwise = |xy: &[(f32, f32)]| {
-            let mut best = f32::INFINITY;
-            for i in 0..xy.len() {
-                for j in (i + 1)..xy.len() {
-                    let dx = xy[i].0 - xy[j].0;
-                    let dy = xy[i].1 - xy[j].1;
-                    let d2 = dx * dx + dy * dy;
-                    if d2 < best {
-                        best = d2;
-                    }
-                }
-            }
-            best.sqrt()
-        };
-
-        let flat_min = min_pairwise(&capital_xy(&flat_pol));
-        let sphere_min = min_pairwise(&capital_xy(&sphere_pol));
-
-        // Direction-only: sphere shouldn't fall below flat. Strict
-        // ≥1.0 with a tiny epsilon allows for f32 round-off when the
-        // two paths happen to agree.
-        assert!(
-            sphere_min + 1e-3 >= flat_min,
-            "sphere min-pairwise province seed distance ({sphere_min:.1} px) regressed below flat \
-             ({flat_min:.1} px); review-impl HIGH-1 fix not in effect"
-        );
-    }
-
-    #[test]
-    fn spherical_view_runs_political_builder_unchanged() {
-        // Civilization stack is mesh-agnostic: political::build only
-        // consumes (seed, centers, neighbors, biomes). On spherical
-        // input it must still produce ≥1 province + ≥1 state.
-        let world = generate(&FlatParams::default());
-        let mut view = build_civ_view(&world, &WorldClimateParams::default());
-        view = augment_with_ocean(view, &world, 64);
-        project_to_sphere(&mut view, &world);
-        let political = crate::political::build(
-            42,
-            &view.centers,
-            &view.neighbors,
-            &view.biomes,
-        );
-        assert!(
-            !political.provinces.is_empty(),
-            "spherical view should still produce provinces"
-        );
-        assert!(
-            !political.states.is_empty(),
-            "spherical view should still produce states"
-        );
-    }
-
-    #[test]
-    fn project_to_sphere_zero_dim_world_is_noop() {
-        // Defensive: a degenerate `width=0` or `height=0` world should
-        // not divide-by-zero. The projection is a no-op so callers can
-        // call it on any FlatWorld without pre-checks.
-        let world = FlatWorld {
-            width: 0,
-            height: 0,
-            plates: vec![],
-            collision_gain: 1.0,
-        };
-        let mut view = CivView {
-            centers: vec![[1.0, 2.0, 3.0]],
-            neighbors: vec![vec![]],
-            biomes: vec![BiomeKind::Ocean],
-            climate: vec![ClimateZone::Tropical],
-            river_flux: vec![0.0],
-            is_coast: vec![false],
-            elevation: vec![0.0],
-            sea_level: 0.5,
-        };
-        project_to_sphere(&mut view, &world);
-        assert_eq!(view.centers[0], [1.0, 2.0, 3.0], "no-op on zero-dim world");
-    }
-
-    #[test]
-    fn civ_bundle_hash_differs_across_seeds() {
-        let world = generate(&FlatParams::default());
-        let a = bundle_civ(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            7,
-            SettlementDensity::Medium,
-            5,
-        );
-        let b = bundle_civ(
-            &world,
-            &WorldClimateParams::default(),
-            32,
-            999,
-            SettlementDensity::Medium,
-            5,
-        );
-        assert_ne!(
-            a.content_hash, b.content_hash,
-            "two distinct seeds should hash differently"
-        );
-    }
-
-    #[test]
-    fn political_png_has_correct_byte_length() {
-        // Ship 8: PNG buffer is `width * height * 3` bytes (RGB).
-        let world = generate(&FlatParams::default());
-        let (_, _, political, _, settlements, _, _) = build_culture(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-            SettlementDensity::Medium,
-            5,
-        );
-        let buf = render_civ_political_png(&world, &political, &settlements);
-        assert_eq!(
-            buf.len(),
-            (world.width as usize) * (world.height as usize) * 3,
-            "PNG buffer should be width*height*3 bytes"
-        );
-    }
-
-    #[test]
-    fn political_png_paints_void_pixels_with_ocean_blue() {
-        // Pixels outside any plate must end up the dark-ocean colour.
-        let world = generate(&FlatParams::default());
-        let (_, _, political, _, settlements, _, _) = build_culture(
-            &world,
-            &WorldClimateParams::default(),
-            64,
-            42,
-            SettlementDensity::Medium,
-            5,
-        );
-        let buf = render_civ_political_png(&world, &political, &settlements);
-        let w = world.width as usize;
-        let h = world.height as usize;
-        let mut void_found = false;
-        for py in 0..h.min(50) {
-            for px in 0..w.min(50) {
-                let fx = px as f32 + 0.5;
-                let fy = py as f32 + 0.5;
-                if world.plates.iter().any(|p| p.contains(fx, fy)) {
-                    continue;
-                }
-                let off = (py * w + px) * 3;
-                assert_eq!(
-                    (buf[off], buf[off + 1], buf[off + 2]),
-                    (12, 22, 48),
-                    "void pixel at ({px},{py}) should be ocean blue (12,22,48)"
-                );
-                void_found = true;
-            }
-        }
-        assert!(
-            void_found,
-            "test world must have ≥1 void pixel in the top-left 50×50 window; \
-             otherwise this test vacuously passes"
-        );
-    }
-
-    #[test]
-    fn civ_svg_export_well_formed_and_contains_features() {
-        // Ship 8: SVG export must start with <?xml + <svg, contain
-        // <polygon> for plates, <circle> for settlements, and <text>
-        // for labels.
-        let world = generate(&FlatParams::default());
-        let (_, mut features, mut political, _, mut settlements, routes_v, mut culture_v) =
-            build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                64,
-                42,
-                SettlementDensity::Medium,
-                5,
-            );
-        apply_synthetic_names(
-            &mut features,
-            &mut political,
-            &mut settlements,
-            &mut culture_v,
-            42,
-        );
-        let svg = render_civ_svg(&world, &political, &settlements, &routes_v);
-        assert!(svg.starts_with("<?xml"), "SVG must start with XML decl");
-        assert!(svg.contains("<svg "), "must contain <svg> root");
-        assert!(svg.contains("<polygon "), "must emit plate polygons");
-        assert!(svg.contains("<circle "), "must emit settlement circles");
-        assert!(svg.contains("<text "), "must emit text labels");
-        assert!(svg.ends_with("</svg>\n"), "must close <svg>");
-    }
-
-    #[test]
-    fn synthetic_names_differ_across_seeds() {
-        // Different seeds should produce different name sets — proves
-        // the seed actually flows into the picker (regression catch).
-        let world = generate(&FlatParams::default());
-
-        let make_bundle = || {
-            let (_, f, p, _, s, _, c) = build_culture(
-                &world,
-                &WorldClimateParams::default(),
-                32,
-                7,
-                SettlementDensity::Medium,
-                5,
-            );
-            (f, p, s, c)
-        };
-        let (mut fa, mut pa, mut sa, mut ca) = make_bundle();
-        let (mut fb, mut pb, mut sb, mut cb) = make_bundle();
-        apply_synthetic_names(&mut fa, &mut pa, &mut sa, &mut ca, 1);
-        apply_synthetic_names(&mut fb, &mut pb, &mut sb, &mut cb, 999);
-        // At least one settlement name must differ between seeds.
-        let differ = sa
-            .iter()
-            .zip(sb.iter())
-            .any(|(a, b)| a.name != b.name);
-        assert!(
-            differ,
-            "two distinct seeds produced identical settlement names; \
-             RNG isn't reaching the picker"
-        );
-    }
-}
